Handle css classes for copy & copied states
In addition to changing the target button's text, the Stimulus classes
`copy` and `copied` can be used to give a visual feedback when copying
to the clipboard.

If defined, the `copied` class is added when the content is copied, while
the `copy` class (once again, if defined) is removed. When the initial text
is restored, so are the CSS classes.