Prefer Clipboard API over execCommand
Document.execCommand is deprecated and will be removed in the future
(see
http://developer.mozilla.org/en-US/docs/Web/API/Document/execCommand for
further details). Clipboard API
(http://developer.mozilla.org/en-US/docs/Web/API/Clipboard_API) is the
replacement to use going forward.

The Clipboard API nets us some improvements over execCommand, namely
that we no longer have to select the text to copy first. This means we
can get rid of the visible input field if we so desire.

It does however also mean we have to abandon Internet Explorer, which
is probably perfectly acceptable.